@@ -18,21 +18,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
-  google_maps_flutter: ^1.1.1
+  google_maps_flutter: ^2.0.6
 
-  geocoder: ^0.2.1
-
+  flutter_geocoder: ^0.2.2-nullsafety
 
   map_pin_picker:
     path: ../
