@@ -7,7 +7,7 @@ repository: https://github.com/pharous-flutter/map_pin_picker
 issue_tracker: https://github.com/pharous-flutter/map_pin_picker/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
